requirements.txt
PyGithub==1.55
requests==2.26.0